+++ impl/src/main/java/uk/ac/manchester/cs/owl/owlapi/OWLLiteralImplPlain.java
@@ -85,7 +85,7 @@ public class OWLLiteralImplPlain extends OWLObjectImpl implements OWLLiteral {
 
     @Override
     protected int hashCode(OWLObject object) {
-        return hash(object.hashIndex(), Stream.of(getDatatype(), specificHash() * 65536, getLang()));
+        return hash(object.hashIndex(), Stream.of(getDatatype(), Integer.valueOf(specificHash() * 65536), getLang()));
     }
 
     private int specificHash() {
