+++ tools/src/main/java/org/semanticweb/owlapi/metrics/AverageAssertedNamedSuperclassCount.java
@@ -69,7 +69,7 @@ public class AverageAssertedNamedSuperclassCount extends DoubleValuedMetric {
                 }
             }
         }
-        return (double) total / count;
+        return Double.valueOf((double) total / count);
     }
 
     @Override
